requirements.txt
anthropic>=0.40.0
browser-use>=0.1.0
playwright>=1.40.0
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0
notion-client>=2.2.0,<3.0.0
PyGithub>=2.1.0
python-dotenv>=1.0.0
jinja2>=3.1.0
pyyaml>=6.0.0
duckduckgo-search>=4.0.0
requests>=2.31.0
